requirements.txt
pydantic
fastapi
mangum
uvicorn
tensorflow-cpu
tensorflow_hub
pandas
numpy
scikit-learn